move tools into submodule
