+++ dodo.py
@@ -2,7 +2,7 @@ import glob
 import os
 import distutils.core
 
-from tools import update_language
+from domdiv.tools import update_language
 
 DOIT_CONFIG = {"default_tasks": ["build"]}
 
@@ -28,7 +28,7 @@ def task_compile_requirements():
 def task_update_languages():
     files = glob.glob("card_db_src/**/*.json") + glob.glob("card_db_src/*.json")
     return {
-        "file_dep": files + ["tools/update_language.py"],
+        "file_dep": files + ["src/domdiv/tools/update_language.py"],
         "actions": [lambda: update_language.main("card_db_src", "src/domdiv/card_db")],
         "targets": [
             os.path.join("src/domdiv/card_db", "/".join(fname.split("/")[1:]))
