+++ pyproject.toml
@@ -1,0 +1,4 @@
+[build-system]
+requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4", "pytest-runner"]
+
+[tools.setuptools_scm]
