+++ src/domdiv/tools/common.py
